# docs/openapi-private.json
{
  "openapi":"3.1.0",
  "info":{"title":"NJKB API NTT — Private Unified Vehicle API (Foundation)","version":"1.0.0-foundation","description":"Planned authenticated private API. Contract foundation only; route is not active or deployed."},
  "servers":[{"url":"https://njkb-api-ntt.elwinmusadi.workers.dev","description":"Target Worker host; /api/v1/vehicle is NOT ACTIVE"}],
  "security":[{"AccessClientId":[],"AccessClientSecret":[]}],
  "x-lifecycle":"pre-deployment-security-gate-ready-production-not-active",
  "x-auth-configuration":{"provider":"Cloudflare Access","client_headers":["CF-Access-Client-Id","CF-Access-Client-Secret"],"origin_assertion_header":"Cf-Access-Jwt-Assertion","issuer":"https://shy-thunder-ffc9.cloudflareaccess.com","audience":"927c4e26c78226a08ecf0a50ed91e74f88587dac5f87ac3a91d1e24eb337e4fa","jwks_url":"https://shy-thunder-ffc9.cloudflareaccess.com/cdn-cgi/access/certs","grant_registry":"AUTH_ACCESS_GRANTS_JSON","canary_grant":{"access_common_name":"d6bb9db60e9cd5ee91327eed387cf2fb.access","principal":"njkb-api-canary","scopes":["vehicle:read","njkb:read"]},"clock_tolerance_seconds":30,"allowed_algorithms":["RS256"],"requirements":["asymmetric signature","issuer","audience","exp","nbf when present","algorithm allowlist"]},
  "paths":{
    "/api/v1/vehicle/{nopol}":{
      "get":{"operationId":"getUnifiedVehicle","summary":"Get authenticated unified vehicle and NJKB data","description":"Implemented for injected-verifier local/integration use. Production provider configuration is not established and the route is not deployed. All query parameters are rejected. Response sections are omitted when the principal lacks their explicit scope.","x-production-active":false,
        "parameters":[{"name":"nopol","in":"path","required":true,"schema":{"type":"string"},"description":"NOPOL normalized using the existing public API rules."}],
        "responses":{
          "200":{"description":"BPAD vehicle found; njkb_status communicates the independent NJKB outcome.","headers":{"Cache-Control":{"schema":{"type":"string","const":"no-store"}},"X-Request-ID":{"schema":{"type":"string","format":"uuid"}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnifiedVehicleResponse"},"example":{"status":"vehicle_found","vehicle_status":"found","njkb_status":"matched","nopol":"ZZ0001ZZ","vehicle":{"category":"MOBIL PENUMPANG","category_raw":"MINIBUS","category_code":"103","brand":"HONDA","brand_code":"167","type":"SYNTHETIC MODEL","type_code":"103167 40649","year":2019,"color":"PUTIH","fuel":"PERTALITE","fuel_code":"01","engine_capacity_cc":1496,"body_modification":"[ --- TIDAK ADA PERUBAHAN --- ]"},"registration":{"bpkb_number":"TEST-BPKB-0001","previous_nopol":"ZZ0000ZZ","stnk_valid_until":"2031-12-31","previous_stnk_date":"2030-12-31","invoice_date":"1900-01-01","plate_color":"PUTIH","usage":"PRIBADI","usage_code":"2"},"tax":{"notice_valid_until":"2030-12-31","previous_pkb_date":"2029-12-31","kohir":"000123"},"owner":{"name":"TEST OWNER","identity_number":"9999999999999999","address":"TEST ADDRESS 123","rt":"001","rw":"002","village":"TEST VILLAGE","district":"TEST DISTRICT","village_code":"99.99.99.9999","district_code":"99.99.99","postal_code":null,"ownership_code":"02","ownership_description":"PEMBELIAN BARU","business_type":null,"business_name":null,"business_number":null,"business_license_number":null},"administrative":{"origin_upt":"TEST UPT","origin_location_code":"010","dealer_code":"016","dealer_name":"TEST DEALER","skum":null,"codes":{"kd_dump":"10","kd_jr":"D4","ke":"1"},"labels":{"k_jr":"SYNTHETIC CLASS"}},"bpad":{"code":"1","status":"success","message":"SYNTHETIC TEST RESPONSE","raw":{"NOPOL":"ZZ0001ZZ","FutureBpadFieldExample":"synthetic-value"}},"njkb":{"value":"150000000.00","weight":"1.050000","dpp_pkb":"157500000.00"},"match":{"method":"exact_code_year_and_brand_type","source_code":"103167 40649"},"source":{"regulation":"Pergub NTT No. 26 Tahun 2025","pdf_page":181,"source_row":"2587"}}}}},
          "400":{"description":"Invalid NOPOL or unsupported query parameter.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrivateErrorResponse"}}}},
          "401":{"description":"Missing, malformed, expired, or unverifiable bearer token.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrivateErrorResponse"}}}},
          "403":{"description":"Authenticated principal lacks required scope.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrivateErrorResponse"}}}},
          "429":{"description":"Per-consumer/application rate limit exceeded.","headers":{"Retry-After":{"schema":{"type":"integer","minimum":1}}},"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrivateErrorResponse"}}}},
          "500":{"description":"Sanitized internal error.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrivateErrorResponse"}}}},
          "502":{"description":"BPAD network, HTTP, JSON, or payload failure.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrivateErrorResponse"}}}},
          "503":{"description":"D1 unavailable.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrivateErrorResponse"}}}},
          "504":{"description":"BPAD timeout.","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PrivateErrorResponse"}}}}
        }
      }
    }
  },
  "components":{
    "securitySchemes":{"AccessClientId":{"type":"apiKey","in":"header","name":"CF-Access-Client-Id","description":"Cloudflare Access service-token client ID supplied by the consumer to Access."},"AccessClientSecret":{"type":"apiKey","in":"header","name":"CF-Access-Client-Secret","description":"Cloudflare Access service-token secret supplied by the consumer to Access. Never logged or stored by the Worker."}},
    "schemas":{
      "NullableString":{"type":["string","null"]},
      "Vehicle":{"type":"object","required":["category","category_raw","category_code","brand","brand_code","type","type_code","year","color","fuel","fuel_code","engine_capacity_cc","body_modification"],"properties":{"category":{"type":"string"},"category_raw":{"type":"string"},"category_code":{"$ref":"#/components/schemas/NullableString"},"brand":{"type":"string"},"brand_code":{"type":"string"},"type":{"type":"string"},"type_code":{"type":"string"},"year":{"type":"integer","minimum":1900,"maximum":2200},"color":{"$ref":"#/components/schemas/NullableString"},"plate_color":{"$ref":"#/components/schemas/NullableString"},"fuel":{"$ref":"#/components/schemas/NullableString"},"fuel_code":{"$ref":"#/components/schemas/NullableString"},"engine_capacity_cc":{"type":["integer","null"]},"chassis_number":{"$ref":"#/components/schemas/NullableString"},"engine_number":{"$ref":"#/components/schemas/NullableString"},"body_modification":{"$ref":"#/components/schemas/NullableString"}}},
      "Registration":{"type":"object","required":["bpkb_number","previous_nopol","stnk_valid_until","previous_stnk_date","invoice_date","plate_color","usage","usage_code"],"properties":{"bpkb_number":{"$ref":"#/components/schemas/NullableString"},"previous_nopol":{"$ref":"#/components/schemas/NullableString"},"stnk_valid_until":{"$ref":"#/components/schemas/NullableString"},"previous_stnk_date":{"$ref":"#/components/schemas/NullableString"},"invoice_date":{"$ref":"#/components/schemas/NullableString"},"plate_color":{"$ref":"#/components/schemas/NullableString"},"usage":{"$ref":"#/components/schemas/NullableString"},"usage_code":{"$ref":"#/components/schemas/NullableString"}}},
      "Tax":{"type":"object","required":["notice_valid_until","previous_pkb_date","kohir"],"properties":{"notice_valid_until":{"$ref":"#/components/schemas/NullableString"},"previous_pkb_date":{"$ref":"#/components/schemas/NullableString"},"kohir":{"$ref":"#/components/schemas/NullableString"}}},
      "Owner":{"type":"object","description":"Phone is intentionally absent because no BPAD phone field is established.","required":["name","identity_number","address","rt","rw","village","district","village_code","district_code","postal_code","ownership_code","ownership_description","business_type","business_name","business_number","business_license_number"],"properties":{"name":{"$ref":"#/components/schemas/NullableString"},"identity_number":{"$ref":"#/components/schemas/NullableString"},"address":{"$ref":"#/components/schemas/NullableString"},"rt":{"$ref":"#/components/schemas/NullableString"},"rw":{"$ref":"#/components/schemas/NullableString"},"village":{"$ref":"#/components/schemas/NullableString"},"district":{"$ref":"#/components/schemas/NullableString"},"village_code":{"$ref":"#/components/schemas/NullableString"},"district_code":{"$ref":"#/components/schemas/NullableString"},"postal_code":{"$ref":"#/components/schemas/NullableString"},"ownership_code":{"$ref":"#/components/schemas/NullableString"},"ownership_description":{"$ref":"#/components/schemas/NullableString"},"business_type":{"$ref":"#/components/schemas/NullableString"},"business_name":{"$ref":"#/components/schemas/NullableString"},"business_number":{"$ref":"#/components/schemas/NullableString"},"business_license_number":{"$ref":"#/components/schemas/NullableString"}}},
      "Administrative":{"type":"object","required":["origin_upt","origin_location_code","dealer_code","dealer_name","skum","codes","labels"],"properties":{"origin_upt":{"$ref":"#/components/schemas/NullableString"},"origin_location_code":{"$ref":"#/components/schemas/NullableString"},"dealer_code":{"$ref":"#/components/schemas/NullableString"},"dealer_name":{"$ref":"#/components/schemas/NullableString"},"skum":{"$ref":"#/components/schemas/NullableString"},"codes":{"type":"object","properties":{"kd_dump":{"$ref":"#/components/schemas/NullableString"},"kd_jr":{"$ref":"#/components/schemas/NullableString"},"ke":{"$ref":"#/components/schemas/NullableString"}}},"labels":{"type":"object","properties":{"k_jr":{"$ref":"#/components/schemas/NullableString"}}}}},
      "BPADRaw":{"type":"object","description":"Exact upstream field names and values. All 56 known fields are strings when present; future unknown JSON-compatible fields are preserved.","properties":{"Alamat":{"type":"string"},"BBM":{"type":"string"},"DEALER":{"type":"string"},"GUNA":{"type":"string"},"IsiCylinder":{"type":"string"},"JENIS":{"type":"string"},"J_USAHA":{"type":"string"},"JenisKendaraan":{"type":"string"},"KD_BBM":{"type":"string"},"KD_DUMP":{"type":"string"},"KD_GUNA":{"type":"string"},"KD_JR":{"type":"string"},"KD_LOKASI_ASAL":{"type":"string"},"KD_MERK":{"type":"string"},"KD_MILIK":{"type":"string"},"KD_POS":{"type":"string"},"KD_TIPE":{"type":"string"},"KE":{"type":"string"},"K_GUNA":{"type":"string"},"K_JR":{"type":"string"},"Kecamatan":{"type":"string"},"Kelurahan":{"type":"string"},"KodeKec":{"type":"string"},"KodeKel":{"type":"string"},"Kohir":{"type":"string"},"MILIK":{"type":"string"},"Merk":{"type":"string"},"NM_DEALER":{"type":"string"},"NM_USAHA":{"type":"string"},"NOPOL":{"type":"string"},"NOPOL_EKS":{"type":"string"},"NO_SIUP":{"type":"string"},"NO_USAHA":{"type":"string"},"NamaPemilik":{"type":"string"},"NoBPKB":{"type":"string"},"NoKTP":{"type":"string"},"NoMesin":{"type":"string"},"NoRangka":{"type":"string"},"Rt":{"type":"string"},"RubahBentuk":{"type":"string"},"Rw":{"type":"string"},"SD_NOTICE":{"type":"string"},"SD_STNK":{"type":"string"},"Skum":{"type":"string"},"TG_FAKTUR":{"type":"string"},"TNKB":{"type":"string"},"T_GUNA":{"type":"string"},"TahunPembuatan":{"type":"string"},"TglPKBLalu":{"type":"string"},"TglSTNKLalu":{"type":"string"},"Type":{"type":"string"},"UPT_ASAL":{"type":"string"},"Warna":{"type":"string"},"kode":{"type":"string"},"pesan":{"type":"string"},"status":{"type":"string"}},"additionalProperties":true},
      "BPADMetadata":{"type":"object","required":["code","status","message"],"properties":{"code":{"$ref":"#/components/schemas/NullableString"},"status":{"$ref":"#/components/schemas/NullableString"},"message":{"$ref":"#/components/schemas/NullableString"},"raw":{"$ref":"#/components/schemas/BPADRaw"}}},
      "NJKB":{"type":["object","null"],"properties":{"value":{"type":["string","null"],"pattern":"^[0-9]+\\.00$"},"weight":{"type":["string","null"],"pattern":"^[0-9]+\\.[0-9]{6}$"},"dpp_pkb":{"type":["string","null"],"pattern":"^[0-9]+\\.00$"}}},
      "Match":{"type":["object","null"],"properties":{"method":{"type":"string"},"source_code":{"type":"string"},"conflicts":{"type":"array","items":{"type":"string"}}}},
      "Source":{"type":["object","null"],"properties":{"regulation":{"type":"string"},"pdf_page":{"type":"integer"},"source_row":{"type":"string"}}},
      "UnifiedVehicleResponse":{"type":"object","required":["status","vehicle_status","njkb_status","nopol","vehicle","bpad"],"properties":{"status":{"const":"vehicle_found"},"vehicle_status":{"const":"found"},"njkb_status":{"enum":["matched","not_found","reference_unavailable","conflict","ambiguous"]},"nopol":{"type":"string"},"vehicle":{"$ref":"#/components/schemas/Vehicle"},"registration":{"$ref":"#/components/schemas/Registration"},"tax":{"$ref":"#/components/schemas/Tax"},"owner":{"$ref":"#/components/schemas/Owner"},"administrative":{"$ref":"#/components/schemas/Administrative"},"bpad":{"$ref":"#/components/schemas/BPADMetadata"},"njkb":{"$ref":"#/components/schemas/NJKB"},"match":{"$ref":"#/components/schemas/Match"},"source":{"$ref":"#/components/schemas/Source"}}},
      "PrivateErrorResponse":{"type":"object","required":["status","error"],"properties":{"status":{"enum":["invalid_request","unauthorized","forbidden","rate_limit_exceeded","upstream_error","error"]},"vehicle_status":{"const":"upstream_error"},"njkb_status":{"type":"null"},"error":{"type":"object","required":["code","message","request_id"],"properties":{"code":{"enum":["invalid_nopol","unsupported_query_parameter","unauthorized","forbidden","rate_limit_exceeded","upstream_error","database_error","internal_error"]},"message":{"type":"string"},"request_id":{"type":"string","format":"uuid"}}}}}
    }
  },
  "x-rate-limit":{"limit":60,"window_seconds":60,"key":"stable project principal","distribution":"per Worker isolate"},
  "x-scope-model":{"vehicle:read":"basic normalized vehicle and BPAD metadata","registration:read":"registration plus sensitive vehicle identifiers","tax:read":"tax section","owner:read":"owner/PII section","njkb:read":"NJKB, match, source","bpad:raw":"exact raw payload","vehicle:full":"explicit bundle of all six granular scopes"},
  "x-privacy":{"cache":"no-store","logging":"raw BPAD and PII are never logged","storage":"raw BPAD is not persisted","phone":"NOT ESTABLISHED"}
}